Title moved to components folder
