+++ components/Title.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+import Typography from "@material-ui/core/Typography";
+
+const Title = styled(Typography)`
+  && {
+    margin: 0;
+    width: 100%;
+    padding-top: 100px;
+    text-align: center;
+    font-weight: bold;
+  }
+`;
+
+export default Title;
