+++ pages/signup.js
@@ -1,21 +1,13 @@
 import React, { useState } from "react";
-import TextField from "@material-ui/core/TextField";
-import MainHeader from "../components/MainHeader";
 import styled from "styled-components";
-import { Typography, Container, Button } from "@material-ui/core";
+
+import TextField from "@material-ui/core/TextField";
 import Grid from "@material-ui/core/Grid";
+import Container from "@material-ui/core/Container";
+import Button from "@material-ui/core/Button";
 
-const CenteredContainer = styled(Container)``;
-
-const Title = styled(Typography)`
-  && {
-    margin: 0;
-    width: 100%;
-    padding-top: 100px;
-    text-align: center;
-    font-weight: bold;
-  }
-`;
+import MainHeader from "../components/MainHeader";
+import Title from "../components/Title";
 
 const StyledTextField = styled(TextField)``;
 
@@ -36,7 +28,7 @@ const Signup = () => {
   const password = useFormInput("");
 
   return (
-    <CenteredContainer>
+    <Container>
       <MainHeader />
       <Title variant="h3">Sign Up</Title>
       <StyledGrid container direction="column" spacing={3} alignItems="center">
@@ -60,14 +52,17 @@ const Signup = () => {
           <StyledButton
             aria-label="Continue"
             onClick={() => {
-              //validar que no exista, registrarme e ir a la welcome page
+              //validar que no exista y registrarme
+              //
+              //
+              //ir a la welcome page
             }}
           >
             Continue
           </StyledButton>
         </Grid>
       </StyledGrid>
-    </CenteredContainer>
+    </Container>
   );
 };
